Implements applyCellWidthParameters prop on TableBodyRow, refactors timeline component models
